Overall fixes

+++ learningplatform-backend/src/main/java/com/manolispapadimitriou/learningplatformbackend/dao/CourseDAO.java
@@ -6,7 +6,7 @@ import lombok.Data;
 @Data
 @JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
 public class CourseDAO {
-    private Long id;
+    private Integer id;
     private String name;
     private String image;
     private String description;
@@ -14,7 +14,7 @@ public class CourseDAO {
     public CourseDAO() {
     }
 
-    public CourseDAO(Long id, String name,String image,String description) {
+    public CourseDAO(Integer id, String name,String image,String description) {
         this.id = id;
         this.name = name;
         this.image = image;
