+++ learningplatform-backend/src/main/java/com/manolispapadimitriou/learningplatformbackend/repository/UserQuizRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface UserQuizRepository extends JpaRepository<UserQuiz, Long> {
+public interface UserQuizRepository extends JpaRepository<UserQuiz, Integer> {
     Optional<UserQuiz> findAllByUserIdAndQuizId(Integer userId, Integer quizId);
     UserQuiz findByUserIdAndQuizId(Integer userId, Integer quizId);
     void deleteAllByUserId(Integer userId);
